move EnumTypeExpected diagnostic to an analyser
